# docs/00-product-baseline/contracts/openapi-v2-p0.yaml
openapi: 3.1.0
info:
  title: Ikaros V2 P0 API
  version: 0.1.0
  description: |
    Machine-readable P0 HTTP contract for Ikaros V2.
    Business semantics are owned by docs/v2/contracts/P0-Command-Query-Event-Catalog.md.
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /admin/storage-providers:
    get:
      tags: [Admin Storage]
      operationId: listStorageProviders
      x-ikaros-contract-id: storage.list-providers
      summary: List configured storage providers
      responses:
        '200':
          description: Storage providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StorageProvider'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Admin Storage]
      operationId: createStorageProvider
      x-ikaros-contract-id: storage.create-provider
      summary: Create a storage provider
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStorageProviderRequest'
      responses:
        '201':
          description: Storage provider created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageProvider'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/storage-providers/{provider_id}/probe:
    post:
      tags: [Admin Storage]
      operationId: probeStorageProvider
      x-ikaros-contract-id: storage.probe-provider
      summary: Probe storage provider connection and read/write capability
      parameters:
        - name: provider_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Probe result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageProviderProbeResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/storage-providers/{provider_id}/credentials:
    post:
      tags: [Admin Storage]
      operationId: replaceStorageProviderCredentials
      x-ikaros-contract-id: storage.replace-provider-credentials
      summary: Replace provider credentials and verify access
      parameters:
        - name: provider_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [access_key_id, secret_access_key]
              properties:
                access_key_id: { type: string, minLength: 1, maxLength: 256, writeOnly: true }
                secret_access_key: { type: string, minLength: 1, maxLength: 512, writeOnly: true }
                session_token: { type: [string, 'null'], maxLength: 2048, writeOnly: true }
      responses:
        '200':
          description: Credential verification result; raw credentials are never returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageProviderProbeResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/storage-providers/{provider_id}/status:
    get:
      tags: [Admin Storage]
      operationId: getStorageProviderStatus
      x-ikaros-contract-id: storage.get-provider-status
      summary: View storage provider health and optional capacity
      parameters:
        - name: provider_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Provider status
          content:
            application/json:
              schema:
                type: object
                required: [provider_id, provider_status, health, checked_at]
                properties:
                  provider_id: { type: string, format: uuid }
                  provider_status: { type: string }
                  health: { $ref: '#/components/schemas/StorageProviderProbeResult' }
                  capacity_bytes: { type: [integer, 'null'], format: int64, minimum: 0 }
                  used_bytes: { type: [integer, 'null'], format: int64, minimum: 0 }
                  checked_at: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /reading/comic-imports:
    post:
      tags: [Reading]
      operationId: createComicImport
      x-ikaros-contract-id: reading.create-comic-import
      summary: Register a supported comic package for import
      parameters:
        - name: X-Ikaros-Actor-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateComicImportRequest' }
      responses:
        '200':
          description: Comic package accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ComicImport' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [Reading]
      operationId: listComicImports
      x-ikaros-contract-id: reading.list-comic-imports
      summary: List comic package imports
      parameters:
        - name: X-Ikaros-Actor-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Comic package imports
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ComicImport' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /reading/comic-imports/{import_id}:
    get:
      tags: [Reading]
      operationId: getComicImport
      x-ikaros-contract-id: reading.get-comic-import
      summary: Get one comic package import
      parameters:
        - name: import_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: X-Ikaros-Actor-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Comic package import
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ComicImport' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /reading/comic-imports/{import_id}/actions/parse:
    post:
      tags: [Reading]
      operationId: parseComicImport
      x-ikaros-contract-id: reading.parse-comic-import
      summary: Parse chapters and page order from a comic package
      parameters:
        - name: import_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: X-Ikaros-Actor-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Comic package parse result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ComicImport' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /reading/comic-imports/{import_id}/entries:
    get:
      tags: [Reading]
      operationId: listComicImportEntries
      x-ikaros-contract-id: reading.list-comic-import-entries
      summary: List discovered comic pages
      parameters:
        - name: import_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: X-Ikaros-Actor-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Discovered chapter and page entries
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ComicImportEntry' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /reading/comic-imports/{import_id}/actions/reorder-pages:
    post:
      tags: [Reading]
      operationId: reorderComicPages
      x-ikaros-contract-id: reading.reorder-comic-pages
      summary: Reorder pages in one discovered comic chapter
      parameters:
        - name: import_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: X-Ikaros-Actor-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReorderComicPagesRequest' }
      responses:
        '200':
          description: Updated comic import
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ComicImport' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /reading/comic-imports/{import_id}/actions/retry-parse:
    post:
      tags: [Reading]
      operationId: retryComicImportParse
      x-ikaros-contract-id: reading.retry-comic-import-parse
      summary: Retry a failed comic package parse
      parameters:
        - name: import_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: X-Ikaros-Actor-Id
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Retried comic package parse result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ComicImport' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /reading/works/{work_id}/bookmarks:
    get:
      tags: [Reading]
      operationId: listReadingBookmarks
      x-ikaros-contract-id: reading.list-bookmarks
      summary: List private reading bookmarks for a work
      parameters:
        - { name: work_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: X-Ikaros-Actor-Id, in: header, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Reading bookmarks, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ReadingBookmark' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /reading/works/{work_id}/editions/{edition_id}/chapters/{chapter_id}/bookmarks:
    post:
      tags: [Reading]
      operationId: createReadingBookmark
      x-ikaros-contract-id: reading.create-bookmark
      summary: Add a private reading bookmark
      parameters:
        - { name: work_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: edition_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: chapter_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: X-Ikaros-Actor-Id, in: header, required: true, schema: { type: string, format: uuid } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ReadingBookmarkRequest' } } } }
      responses:
        '200': { description: Created bookmark, content: { application/json: { schema: { $ref: '#/components/schemas/ReadingBookmark' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /reading/bookmarks/{bookmark_id}:
    delete:
      tags: [Reading]
      operationId: deleteReadingBookmark
      x-ikaros-contract-id: reading.delete-bookmark
      summary: Remove a private reading bookmark
      parameters:
        - { name: bookmark_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: X-Ikaros-Actor-Id, in: header, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Bookmark removed }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-or-jwt
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 200
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authorization denied
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found or not visible
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Business or idempotency conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        request_id: { type: string }
        trace_id: { type: string }
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
...
    CreateComicImportRequest:
      type: object
      required: [attachment_id]
      additionalProperties: false
      properties:
        attachment_id: { type: string, format: uuid }
        title: { type: [string, 'null'] }
        language: { type: [string, 'null'] }
...
    ComicImport:
      type: object
      required: [id, source_attachment_id, work_id, edition_id, status, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        source_attachment_id: { type: string, format: uuid }
        work_id: { type: string, format: uuid }
        edition_id: { type: string, format: uuid }
        status: { type: string, enum: [ACCEPTED, PARSING, SUCCEEDED, FAILED] }
        error_code: { type: [string, 'null'] }
        error_message: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ComicImportEntry:
      type: object
      required: [id, import_id, chapter_key, entry_name, page_order, page_role]
      properties:
        id: { type: string, format: uuid }
        import_id: { type: string, format: uuid }
        chapter_key: { type: string }
        entry_name: { type: string }
        page_order: { type: integer, minimum: 0 }
        page_role: { type: string, enum: [NORMAL, COVER] }
...
    ReadingBookmarkRequest:
      type: object
      required: [locator_kind, locator_value]
      properties:
        locator_kind: { type: string, minLength: 1, maxLength: 32 }
        locator_value: { type: string, minLength: 1, maxLength: 1024 }
        content_version: { type: [string, 'null'] }
        label: { type: [string, 'null'] }
    ReadingBookmark:
      type: object
      required: [id, work_id, edition_id, chapter_id, locator_kind, locator_value, created_at]
      properties:
        id: { type: string, format: uuid }
        work_id: { type: string, format: uuid }
        edition_id: { type: string, format: uuid }
        chapter_id: { type: string, format: uuid }
        locator_kind: { type: string }
        locator_value: { type: string }
        content_version: { type: [string, 'null'] }
        label: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
    ReorderComicPagesRequest:
      type: object
      required: [chapter_key, entry_ids]
      additionalProperties: false
      properties:
        chapter_key: { type: string, minLength: 1 }
        entry_ids:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, format: uuid }
...
    StorageProvider:
      type: object
      required: [id, provider_key, provider_type, display_name, tier, enabled, drain_status, version]
      properties:
        id: { type: string, format: uuid }
        provider_key: { type: string }
        provider_type: { type: string }
        display_name: { type: string }
        tier:
          type: string
          enum: [HOT, WARM, COLD, ARCHIVE]
        enabled: { type: boolean }
        drain_status:
          type: string
          enum: [NORMAL, DRAINING, DRAINED]
        capabilities:
          type: object
          additionalProperties: true
        version: { type: integer, format: int64 }
    StorageProviderProbeResult:
      type: object
      required: [provider_id, status, connection, read, write, checked_at]
      properties:
        provider_id: { type: string, format: uuid }
        status: { type: string, enum: [HEALTHY, DEGRADED, FAILED, UNSUPPORTED] }
        connection: { type: boolean }
        read: { type: boolean }
        write: { type: boolean }
        checked_at: { type: string, format: date-time }
        error_code: { type: [string, 'null'] }
    CreateStorageProviderRequest:
      type: object
      required: [provider_key, provider_type, display_name, tier, capabilities]
      properties:
        provider_key: { type: string, minLength: 1 }
        provider_type: { type: string, minLength: 1 }
        display_name: { type: string, minLength: 1 }
        credential_ref:
          type: [string, 'null']
          description: Secret reference only; plaintext credentials are forbidden.
        tier:
          type: string
          enum: [HOT, WARM, COLD, ARCHIVE]
        capabilities:
          type: object
          additionalProperties: true
...